[ramda] split tests to separate modules - functions beginning with 'b' - 'c' (#38251)
* move 'binary' tests to separate file

* move 'bind' tests to separate file

* move 'both' tests to separate file

* move 'chain' tests to separate file

* move 'clamp' tests to separate file

* move 'clone' tests to separate file

* move 'comparator' tests to separate file

* move 'complement' tests to separate file

* move 'compose' tests to separate file

* move 'composeK' tests to separate file

* move 'composeP' tests to separate file

* move 'composeWith' tests to separate file

* move 'concat' tests to separate file

* move 'cond' tests to separate file

* move 'cond' tests to separate file

* move 'construct' tests to separate file

* move 'constructN' tests to separate file

* move 'contains' tests to separate file

* move 'converge' tests to separate file
