+++ types/ramda/test/bind-tests.ts
@@ -1,0 +1,5 @@
+import * as R from 'ramda';
+
+() => {
+  R.bind(console.log, console);
+};
